requirements.txt
blessings==1.6
bpython==0.15
curtsies==0.2.6
greenlet==0.4.9
Pygments==2.1.3
requests==2.10.0
scapy-python3==0.18
six==1.10.0
wcwidth==0.1.6
